+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/samplelist/FailedQCSamplesView.groovy
@@ -2,12 +2,14 @@ package life.qbic.portal.sampletracking.components.projectoverview.samplelist
 
 
 import com.vaadin.data.provider.DataProvider
-import com.vaadin.ui.Grid
-import com.vaadin.ui.VerticalLayout
+import com.vaadin.icons.VaadinIcons
+import com.vaadin.shared.ui.grid.HeightMode
+import com.vaadin.ui.*
 import life.qbic.business.samples.info.GetSamplesInfoOutput
 import life.qbic.datamodel.samples.Status
 import life.qbic.portal.sampletracking.communication.notification.NotificationService
-import life.qbic.portal.sampletracking.components.NotificationHandler
+import life.qbic.portal.sampletracking.components.projectoverview.visibility.VisibilityChangeEvent
+import life.qbic.portal.sampletracking.components.projectoverview.visibility.VisibilityChangeListener
 
 /**
  * <b>Shows the failed QC samples </b>
@@ -19,6 +21,7 @@ class FailedQCSamplesView extends VerticalLayout {
     private final Presenter presenter
 
     private Grid<Sample> samplesGrid
+    private final List<VisibilityChangeListener> visibilityChangeListeners = []
 
     FailedQCSamplesView(NotificationService notificationService) {
         this.viewModel = new ViewModel()
@@ -29,18 +32,60 @@ class FailedQCSamplesView extends VerticalLayout {
     private void initLayout() {
         this.setMargin(false)
         createSamplesGrid()
+        HorizontalLayout buttonLayout = setupCloseButtonLayout()
 
-        this.addComponent(samplesGrid)
+        this.addComponents(buttonLayout,samplesGrid)
+    }
+
+    void addVisibilityChangeListener(VisibilityChangeListener listener){
+        visibilityChangeListeners.add(listener)
+    }
+
+    private HorizontalLayout setupCloseButtonLayout() {
+        Button closeButton = new Button("Hide", VaadinIcons.CLOSE_CIRCLE)
+        closeButton.addClickListener({
+            this.setVisible(false)
+        })
+
+        HorizontalLayout buttonLayout = new HorizontalLayout()
+        buttonLayout.addComponent(closeButton)
+        buttonLayout.setComponentAlignment(closeButton, Alignment.TOP_LEFT)
+
+        return buttonLayout
+    }
+
+    @Override
+    void setVisible(boolean visible) {
+        boolean currentlyVisible = isVisible()
+        if (currentlyVisible != visible) {
+            VisibilityChangeEvent changeEvent = new VisibilityChangeEvent(this, this.isVisible(), visible)
+            fireVisibilityChangeEvent(changeEvent)
+        }
+        super.setVisible(visible)
+    }
+
+    private void fireVisibilityChangeEvent(VisibilityChangeEvent changeEvent) {
+        visibilityChangeListeners.forEach({ it.visibilityChanged(changeEvent) })
+    }
+
+    /**
+     * Resets the view to its initial state.
+     * @since 1.0.0
+     */
+    void reset() {
+        if (viewModel.samples.size() > 0) {
+            viewModel.samples.clear()
+        }
     }
 
     private void createSamplesGrid() {
 
         this.samplesGrid = new Grid<>()
-        samplesGrid.setCaption("Samples that failed quality control")
-        samplesGrid.addColumn(Sample::getCode).setCaption("Sample Code").setId("SampleCode")
+        samplesGrid.addColumn(Sample::getCode).setCaption("Failed QC Sample Code").setId("SampleCode")
         samplesGrid.addColumn(Sample::getTitle).setCaption("Sample Title").setId("SampleTitle")
         samplesGrid.setSelectionMode(Grid.SelectionMode.NONE)
         samplesGrid.setDataProvider(DataProvider.ofCollection(viewModel.getSamples()))
+        samplesGrid.setHeightMode(HeightMode.ROW)
     }
 
     GetSamplesInfoOutput getPresenter() {